+++ examples/dockerhostcli.py
@@ -23,10 +23,16 @@ def testTopo():
 
     info("*** Adding hosts\n")
     h1 = net.addDockerHost(
-        "h1", dimage="dev_test", ip="10.0.0.1/24", cpuset_cpus="0", cpu_quota=25000
+        "h1",
+        dimage="dev_test",
+        ip="10.0.0.1/24",
+        docker_args={"cpuset_cpus": "0", "nano_cpus": int(1e8)},
     )
     h2 = net.addDockerHost(
-        "h2", dimage="dev_test", ip="10.0.0.2/24", cpuset_cpus="1", cpu_quota=25000
+        "h2",
+        dimage="dev_test",
+        ip="10.0.0.2/24",
+        docker_args={"cpuset_cpus": "0", "nano_cpus": int(1e8)},
     )
     h3 = net.addHost("h3", ip="10.0.0.3/24")
     h4 = net.addHost("h4", ip="10.0.0.4/24")